数据总数改为int类型

+++ src/main/java/org/parts/response/EasyUiPageVo.java
@@ -10,16 +10,15 @@ import java.util.List;
 
 /**
  * easyui分页统一返回实体类
- * 
- * @author ltl
  *
+ * @author ltl
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class EasyUiPageVo<T> implements Serializable {
-	private static final long serialVersionUID = -5123560345302553119L;
-	private List<T> rows;
-	private Long total;
+    private static final long serialVersionUID = -5123560345302553119L;
+    private List<T> rows;
+    private Integer total;
 }
